Bug fixes
Midair Mantle Anim State transition caused a problem. I fixed that good. (had to use a greater than or equal to node, NOT a mere greater than node).

Fixed the Ledge Warping issue by adding a condition in the Update Tick by adding a clause to the condition statement to check feet only if in the midair state.